split Viewer feedback economics model
